+++ ingester/src/data/triggers/test.rs
@@ -1,4 +1,5 @@
 use std::sync::Arc;
+
 use tokio::sync::{Barrier, Mutex};
 
 #[derive(Debug, Clone)]
